requirements.txt
PyYAML
pyyaml
pre-commit
mypy
black
isort
pylint
coverage
types-PyYAML
types-requests
